Fix issue with utilized Gradle version when working with multi module projects
For the root project, all would be well in the world. However, for any subproject, it would fail to find the Gradle wrapper and thus use the presently configured 7.6 version. If the subproject happened to utilize removed functionality between the version the project used and 7.6, this would have resulted in a build error