Map gradle dependency and configuration attributes to openrewirte attributes. (#44)
Pairs with: https://github.com/openrewrite/rewrite/commit/55c33b41bec930b4dfb55d8de0747e3fc0815031

Also restore the ability to publish this module without first waiting for a release of rewrite.

We're supposed to be able to build and publish this from snapshots so that, even when this depends on rewrite-core changes, a release version of this can be available before the corresponding changes in rewrite-core are available in a release build.